# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/v1/ecommerce/signup": {
      "post": {
        "description": "user can signup by giving their details",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["User"],
        "summary": "User Signup",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
...
}